Calculate day-level summaries from the palmsplus dataset

Build a days dataset by summarising palmsplus by day and person (identifier). Not all variables in palmsplus are summarised, only those specified using palms_add_field with the argument domain_field = TRUE. By default, a duration field is added (e.g., the total minutes per day).

All data are summarised by default. However, additional aggragation domains can be specified using palms_add_domain before building palmsplus. Domains are a subset of data, such as during school time. All domain_field variables will be summarised for each domain seperatly.

palms_build_days(data)

Arguments

data
The palmsplus data obtained from palms_build_palmsplus.

Value

A table summarised by day.

Examples

data("palms")

# Add a field, and make it a domain_field
palms_add_field("mvpa", "activityintensity > 1", TRUE)

palms_add_domain("walking", "tripmot == 3")
palmsplus <- palms_build_palmsplus(palms)

#> Error in mutate_impl(.data, dots): Evaluation error: object 'home.buffer' not found.

# This will have 'walking' and 'total' domains
# Each domain will have 'duration' and 'mvpa' fields
days <- palms_build_days(palmsplus)

#> Error in data$total <- 1: object 'palmsplus' not found